Broke out common functionality of service structs to protocol.
